+++ npm-publish-gradle-plugin/src/main/kotlin/config/packages.kt
@@ -12,17 +12,13 @@ internal fun ProjectEnhancer.configure(packages: NpmPackages) {
     val assTask = tasks.register(assembleTaskName(it.name), NpmAssembleTask::class.java) { task ->
       task.description = "Assembles ${it.name} package."
       task.`package`.set(it)
-    }.also { task ->
-      info { "Registered [${task.name}] NpmAssembleTask for [${it.name}] NpmPackage" }
-    }
+    }.also { task -> info { "Registered [${task.name}] NpmAssembleTask for [${it.name}] NpmPackage" } }
     tasks.register(packTaskName(it.name), NpmPackTask::class.java) { task ->
       task.description = "Packs ${it.name} package."
       task.dry.set(extension.dry)
       task.packageDir.set(assTask.flatMap(NpmAssembleTask::destinationDir))
       task.nodeHome.set(extension.nodeHome)
-    }.also { task ->
-      info { "Registered [${task.name}] NpmPackTask for [${it.name}] NpmPackage" }
-    }
+    }.also { task -> info { "Registered [${task.name}] NpmPackTask for [${it.name}] NpmPackage" } }
     extension.registries.names.forEach { registryName ->
       registerPublishTask(it.name, registryName)
     }